abstraction of all webview html
